+++ src/News/UseCase/Create/Command.php
@@ -10,7 +10,8 @@ readonly class Command
 {
     public function __construct(
         public string $title,
-        public string $text
+        public string $text,
+        public ?string $imageUrl = null
     ) {
         $this->validate();
     }
@@ -26,6 +27,10 @@ readonly class Command
 
         if ('' === trim($this->text)) {
             throw new InvalidArgumentException('news text cannot be empty');
+        }
+
+        if (null !== $this->imageUrl && '' === trim($this->imageUrl)) {
+            throw new InvalidArgumentException('news image url cannot be empty');
         }
     }
 }